+++ src/main/java/forestry/api/apiculture/ForestryBeeSpecies.java
@@ -71,4 +71,6 @@ public class ForestryBeeSpecies {
 	public static final ResourceLocation SPITEFUL = forestry("bee_spiteful");
 	public static final ResourceLocation SEETHING = forestry("bee_seething");
 	public static final ResourceLocation WARPED = forestry("bee_warped");
+	public static final ResourceLocation ZOMBIFIED = forestry("bee_zombified");
+	public static final ResourceLocation SCULK = forestry("bee_sculk");
 }