+++ src/main/java/forestry/api/genetics/alleles/ForestryAlleles.java
@@ -50,6 +50,7 @@ public class ForestryAlleles {
 	public static final IIntegerAllele LIFESPAN_LONG = REGISTRY.intAllele(50);
 	public static final IIntegerAllele LIFESPAN_LONGER = REGISTRY.intAllele(60);
 	public static final IIntegerAllele LIFESPAN_LONGEST = REGISTRY.intAllele(70);
+	public static final IIntegerAllele LIFESPAN_IMMORTAL = REGISTRY.intAllele(Integer.MAX_VALUE);
 
 	public static final List<IIntegerAllele> DEFAULT_LIFESPANS = List.of(LIFESPAN_SHORTEST, LIFESPAN_SHORTER, LIFESPAN_SHORT, LIFESPAN_SHORTENED, LIFESPAN_NORMAL, LIFESPAN_ELONGATED, LIFESPAN_LONG, LIFESPAN_LONGER, LIFESPAN_LONGEST);
 
@@ -131,6 +132,7 @@ public class ForestryAlleles {
 	public static final IRegistryAllele<IFlowerType> FLOWER_TYPE_PHOTOSYNTHESIS = REGISTRY.registryAllele(ForestryFlowerTypes.PHOTOSYNTHESIS, BeeChromosomes.FLOWER_TYPE);
 	public static final IRegistryAllele<IFlowerType> FLOWER_TYPE_SEA = REGISTRY.registryAllele(ForestryFlowerTypes.SEA, BeeChromosomes.FLOWER_TYPE);
 	public static final IRegistryAllele<IFlowerType> FLOWER_TYPE_CORAL = REGISTRY.registryAllele(ForestryFlowerTypes.CORAL, BeeChromosomes.FLOWER_TYPE);
+	public static final IRegistryAllele<IFlowerType> FLOWER_TYPE_SCULK = REGISTRY.registryAllele(ForestryFlowerTypes.SCULK, BeeChromosomes.FLOWER_TYPE);
 
 	// Bee Effect
 	public static final IRegistryAllele<IBeeEffect> EFFECT_NONE = REGISTRY.registryAllele(ForestryBeeEffects.NONE, BeeChromosomes.EFFECT);
@@ -158,6 +160,7 @@ public class ForestryAlleles {
 	public static final IRegistryAllele<IBeeEffect> EFFECT_GUARDIAN = REGISTRY.registryAllele(ForestryBeeEffects.GUARDIAN, BeeChromosomes.EFFECT);
 	public static final IRegistryAllele<IBeeEffect> EFFECT_PHASING = REGISTRY.registryAllele(ForestryBeeEffects.PHASING, BeeChromosomes.EFFECT);
 	public static final IRegistryAllele<IBeeEffect> EFFECT_ASCENSION = REGISTRY.registryAllele(ForestryBeeEffects.ASCENSION, BeeChromosomes.EFFECT);
+	public static final IRegistryAllele<IBeeEffect> EFFECT_SCULK = REGISTRY.registryAllele(ForestryBeeEffects.SCULK, BeeChromosomes.EFFECT);
 	//public static final IRegistryAllele<IBeeEffect> EFFECT_PATRIOTIC = Todos.todo();
 
 	// Speed