+++ src/main/java/forestry/core/data/ForestryBlockTagsProvider.java
@@ -174,6 +174,7 @@ public final class ForestryBlockTagsProvider {
 		tags.tag(ForestryTags.Blocks.CAVE_FLOWERS).add(Blocks.CAVE_VINES, Blocks.CAVE_VINES_PLANT, Blocks.SPORE_BLOSSOM);
 		tags.tag(ForestryTags.Blocks.SEA_FLOWERS).add(Blocks.KELP, Blocks.KELP_PLANT);
 		tags.tag(ForestryTags.Blocks.CORAL_FLOWERS).addTags(BlockTags.CORALS).addTags(BlockTags.WALL_CORALS);
+		tags.tag(ForestryTags.Blocks.SCULK_FLOWERS).add(Blocks.SCULK_SHRIEKER, Blocks.SCULK_CATALYST, Blocks.SCULK_SENSOR);
 
 		tags.tag(ForestryTags.Blocks.PLANTABLE_FLOWERS)
 				.addTag(BlockTags.FLOWERS)